backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.35.0
faiss-cpu==1.8.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
numpy>=1.21.0
tiktoken==0.7.0
pypdf==4.2.0
python-docx==1.1.2
google-generativeai
